refactor: 旧SQL方式を削除しバルク方式に完全移行
BulkRecommendUpdaterのロジックをRecommendUpdaterに統合し、
旧SQL方式の15メソッドを削除。比較テストも役目を終えたため削除。
